Can't remember what I changed but I felt that this was a good moment to commit

+++ db/spending.sql
@@ -17,7 +17,7 @@ CREATE TABLE accounts (
   id SERIAL8 PRIMARY KEY,
   first_name VARCHAR(255),
   sur_name VARCHAR(255),
-  budget INT8
+  budget FLOAT
 );
 
 CREATE TABLE transactions (
